requirements.txt
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
